Updated to main GUI splits.
There is an issue when trying to display just the PCB. If layer table is
enabled then everything works fine, but if disabled then the PCB will
not be drawn.

This has something to do with
```js
let width   = document.getElementById(layerID).clientWidth * 2;
```

For some reason this reports a width of 0.