Resolved issue with layer table and PCB split
When layer table is removed, need to call setLayerSplit in order to
resize the other splits correctly. By not calling setLayertSplit when
disabling the layer table, the GUI will not assign a a parent width
to the Canvas element (both layer and canvas elements are tied together)
which then results in the GUI to assign a 0 to the calculated size
which is then cascaded to each layer canvas resulting in them not being
displayed.