Modified makefile for easier debugging.
While uglifying CSS, JS, and HTML makes the resulting program smaller,
this also makes it very hard to debug. So for now all uglify commands
are commented out so that debugging can be done.

Ideally what needs to happen is a if/else needs to be implemented
for each make rule so that if release then make will uglify and if
normal build then uglify is not used.