Added multi selection to BOM table
BOM table now allows pressing CTL+Click to select multiple rows.

A feature of Firefox requires that an additional CSS parameter be added
to a table row element, ```-moz-user-select: none;```, this prevents a
border form appearing around a table element. This behavior I found
only in Firefox. Solution was found on [StackOverflow](https://stackoverflow.com/questions/5067644/html-table-when-i-ctrlclick-the-border-of-the-cell-appears)

This commit starts the implementation for Issue #73.