Fixed bug when clicking module
Was not passing event argument down the chain. After clicking a part, at
some point the BOM row handler vent will trigger and this event expects
the event object as an argument to check if CTL is pressed. The way the
function was defined the evnet handler was not passed in.