Changed part pad rule & elongation specification.
Modified part pads to remove THT/SMD. Only rectangle based pads are now
supported. This simplifies part pad definitions.

Also adjusted elongation too be an integer instead of a real number.
Elongation is a percent that ranges between 0 - 100 and was initially
declared as a real number but should of been a integer since it
represents a percentage.